Staging latest updates (#4)
* fix: theme flickering on load

* refactor: removed old carousel components

* refactor: removed unneeded redux

* refactor: new theme switcher icons

* chore: added calendar icon to button

* chore: sync gallery image with card
