+++ package.json
@@ -12,8 +12,6 @@
     "node": ">=18"
   },
   "dependencies": {
-    "@heroicons/react": "^2.0.18",
-    "@reduxjs/toolkit": "^1.9.5",
     "@tailwindcss/typography": "^0.5.9",
     "@types/node": "20.2.5",
     "@types/react": "18.2.8",
@@ -31,9 +29,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-icons": "^4.9.0",
-    "react-redux": "^8.0.7",
     "react-responsive-carousel": "^3.2.23",
-    "redux": "^4.2.1",
     "sass": "^1.62.1",
     "tailwindcss": "3.3.2",
     "the-new-css-reset": "^1.8.4",
