--- src/components/carousel/carousel.module.scss
@@ -1,69 +1,0 @@
-.carousel {
-  @apply flex flex-col gap-[2px] h-full w-[100svw];
-  .carousel__figure {
-    @apply h-full bg-center bg-cover max-lg:bg-contain bg-no-repeat rounded-t-sm;
-    .carousel__figure_header {
-      @apply text-4xl leading-8 font-semibold mb-12 text-slate-800;
-    }
-  }
-  .carousel__slider__container {
-    @apply relative flex gap-1 scroll-smooth 
-      p-0 m-0;
-  }
-  .carousel__slider {
-    @apply relative;
-  }
-}
-
-.nav_button {
-  @apply hover:bg-blue-900/75 text-white w-10 h-full text-center opacity-75 hover:opacity-100 disabled:opacity-25 disabled:cursor-not-allowed 
-    z-10 p-0 m-0 transition-all ease-in-out duration-150;
-  svg {
-    @apply h-12 w-20 -ml-5;
-  }
-  span {
-    @apply sr-only;
-  }
-}
-@media ((hover: hover) and (pointer: fine)) {
-  .carousel__slider_nav {
-    @apply visible;
-  }
-  .carousel__slider__container {
-    @apply snap-x snap-mandatory z-0;
-  }
-}
-
-@media not ((hover: hover) and (pointer: fine)) {
-  .carousel__figure {
-    @apply hidden;
-  }
-  .carousel__slider_nav {
-    @apply hidden;
-  }
-  .carousel__slider__container {
-    @apply touch-pan-x overflow-x-auto;
-  }
-}
-
-.carousel__image {
-  @apply text-center relative h-32 max-lg:h-16 snap-start cursor-pointer
-  transition hover:scale-95;
-  input {
-    @apply hidden;
-  }
-  label {
-    @apply h-full w-full aspect-square block 
-      bg-origin-padding bg-left-top bg-cover bg-no-repeat 
-      z-0 rounded-b-md;
-  }
-  div {
-    @apply h-full w-full aspect-square block absolute top-0 left-0 
-      transition-opacity duration-150 hover:opacity-100 
-      bg-slate-800/75 z-10;
-    h3 {
-      @apply text-white py-6 px-3 mx-auto text-xl 
-        max-lg:py-3 max-lg:px-0 max-lg:text-xs;
-    }
-  }
-}
