--- src/components/img-carousel/img-carousel.module.scss
@@ -1,62 +1,0 @@
-.carousel {
-  @apply flex flex-col gap-[2px] m-0 h-80 w-80 overflow-hidden;
-  .carousel__figure {
-    @apply h-full bg-center bg-cover;
-    .carousel__figure_header {
-      @apply text-4xl leading-8 font-semibold mb-12 text-slate-800;
-    }
-  }
-  .carousel__container {
-    @apply relative flex gap-1 
-      overflow-hidden max-sm:overflow-x-auto 
-      scroll-smooth snap-x snap-mandatory touch-auto
-      z-0 p-0 m-0;
-  }
-}
-.nav {
-  @apply relative overflow-hidden;
-  .navButtons {
-    @apply flex justify-between absolute top-0 left-0 w-full h-full items-center;
-  }
-}
-.nav_button {
-  @apply hover:bg-stone-900/20 
-    text-white flex items-center justify-center rounded-sm
-    w-11 h-full p-0 m-0 z-[1]
-    opacity-75 disabled:opacity-50 
-    disabled:cursor-not-allowed 
-    transition-all ease-in-out duration-300;
-  svg {
-    @apply h-12 w-12;
-  }
-  span {
-    @apply sr-only;
-  }
-}
-
-.carousel__image {
-  @apply text-center relative w-full h-full snap-start;
-  input {
-    @apply hidden;
-  }
-  label {
-    @apply w-[320px] aspect-square block 
-      bg-origin-padding bg-left-top bg-cover bg-no-repeat 
-      z-0 cursor-zoom-in;
-  }
-}
-
-.dialog {
-  @apply flex-col justify-center items-center 
-    overflow-x-hidden overflow-y-auto 
-    fixed inset-0 z-50 outline-none focus:outline-none
-    w-full h-full p-0 bg-slate-900;
-  .button_close {
-    @apply absolute 
-      right-5 top-5 h-10 w-10 
-      rounded-full 
-      bg-slate-500 hover:bg-slate-300 
-      flex justify-center items-center 
-      font-extrabold text-slate-300 hover:text-slate-950;
-  }
-}
